Fix trip-info and trip-cost.
