+++ src/view/trip-cost.js
@@ -1,22 +1,19 @@
-import { MOCK_EVENTS } from '../mock/events';
 import {createElement} from '../utils.js';
 
-const reducer = (accumulator, currentValue) => accumulator + currentValue;
-const price = MOCK_EVENTS.map((event) => event['base_price']).reduce(reducer);
-
-const createTripCostTemplate = () => (
+const createTripCostTemplate = (price) => (
   `<p class="trip-info__cost">
     Total: €&nbsp;<span class="trip-info__cost-value">${price}</span>
   </p>`
 );
 
 export default class TripCost {
-  constructor() {
+  constructor(price) {
     this._element = null;
+    this._price = price;
   }
 
   getTemplate() {
-    return createTripCostTemplate();
+    return createTripCostTemplate(this._price);
   }
 
   getElement() {
